utils/timestamp: Refactor class to allow operations
The change was needed as the old version only provided static methods
which returned either a unix timestamp or the string version.

The new version of this object allows operations with timestamps.
It adds the get_difference() function to simplify the operations between
timestamps and returns the result in the right resolution.